refactor(web-analytics): relocate presentation to canonical paths
Field-test artifact: ran the fixed `hogli product:isolate:move web_analytics`
end-to-end. Moves the backend/api/ subpackage to presentation/views/ and
serializers.py to presentation/serializers.py, rewriting every reference
repo-wide (imports, bare-package imports via __init__, and @patch strings).

Structure move only — isolation is NOT enabled (no facade, no tach interface,
no contract-check). web_analytics has test-infrastructure coupling (core hogql
tests monkeypatch its runner modules and import its test bases) that blocks a
clean facade isolation without relocating core tests.
