--- products/web_analytics/backend/api/__init__.py
@@ -1,3 +1,0 @@
-from .api import WebAnalyticsViewSet
-
-__all__ = ["WebAnalyticsViewSet"]
