+++ products/web_analytics/backend/presentation/__init__.py
@@ -1,0 +1,1 @@
+"""HTTP presentation layer of the web_analytics product (DRF viewsets and serializers)."""
