+++ products/web_analytics/backend/presentation/views/__init__.py
@@ -1,0 +1,3 @@
+from products.web_analytics.backend.presentation.views.api import WebAnalyticsViewSet
+
+__all__ = ["WebAnalyticsViewSet"]
