+++ products/web_analytics/backend/presentation/views/heatmaps_utils.py
